rename test file, delete old
